+++ apps/backend/src/main/java/io/github/artsobol/kurkod/model/request/chicken/ChickenPatchRequest.java
@@ -10,7 +10,7 @@ import java.time.LocalDate;
 @AllArgsConstructor
 @NoArgsConstructor
 public class ChickenPatchRequest {
-    private short weight;
+    private Integer weight;
 
     private LocalDate birthDate;
